refactor(api): organize application slices
Introduce an application persistence abstraction and move use cases into slice folders with request and response contracts colocated by feature. Scope consumption and resource access through the current user context while preserving the existing HTTP surface.
