+++ .gitignore
@@ -11,6 +11,7 @@ node_modules
 *.njsproj
 *.sln
 *.sw?
+*.lscache
 
 # env
 .env.local
@@ -39,4 +40,3 @@ lerna-debug.log*
 # tools
 .fallow
 .codacy
-
