+++ apps/api/src/Api/Endpoints/ConsumptionsEndpoints.cs
@@ -1,7 +1,13 @@
 using Kijk.Api.Extensions;
 using Kijk.Api.Models;
-using Kijk.Application.Consumptions;
+using Kijk.Application.Consumptions.Create;
+using Kijk.Application.Consumptions.Delete;
+using Kijk.Application.Consumptions.GetById;
+using Kijk.Application.Consumptions.GetByYearMonth;
+using Kijk.Application.Consumptions.GetStats;
+using Kijk.Application.Consumptions.GetYears;
 using Kijk.Application.Consumptions.Shared;
+using Kijk.Application.Consumptions.Update;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
