+++ apps/api/src/Application/Consumptions/GetById/GetByIdConsumption.cs
@@ -1,21 +1,21 @@
+using Kijk.Application.Abstractions.Persistence;
 using Kijk.Application.Consumptions.Shared;
-using Kijk.Infrastructure.Persistence;
 using Kijk.Shared;
 using Microsoft.Extensions.Logging;
 
-namespace Kijk.Application.Consumptions;
+namespace Kijk.Application.Consumptions.GetById;
 
 /// <summary>
 /// Handler for getting consumption by id.
 /// </summary>
-public class GetByIdConsumptionHandler(AppDbContext dbContext, ILogger<GetByIdConsumptionHandler> logger) : IHandler
+public class GetByIdConsumptionHandler(IAppDbContext dbContext, CurrentUser currentUser, ILogger<GetByIdConsumptionHandler> logger) : IHandler
 {
     public async Task<Result<ConsumptionResponse>> GetByIdAsync(Guid id, CancellationToken cancellationToken)
     {
         var entity = await dbContext.Consumptions
             .AsNoTracking()
             .Include(x => x.Resource)
-            .Where(x => x.Id == id)
+            .Where(x => x.Id == id && x.HouseholdId == currentUser.ActiveHouseholdId)
             .Select(x => new ConsumptionResponse(
                 x.Id,
                 x.Name,
