+++ apps/api/src/Application/Consumptions/GetYears/Response.cs
@@ -1,0 +1,3 @@
+namespace Kijk.Application.Consumptions.GetYears;
+
+public record GetYearsConsumptionQueryResponse(IList<int> Years);
