+++ apps/api/src/Application/Consumptions/Update/Request.cs
@@ -1,0 +1,8 @@
+using NetEscapades.EnumGenerators;
+
+namespace Kijk.Application.Consumptions.Update;
+
+public record UpdateConsumptionRequest(string? Name, decimal? Value, UpdateConsumptionValueTypes ValueType, Guid? ResourceId, DateTime? Date);
+
+[EnumExtensions]
+public enum UpdateConsumptionValueTypes { Absolute, Relative };
