+++ apps/api/src/Application/DependencyInjection.cs
@@ -1,7 +1,6 @@
 using System.Collections.Immutable;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Serilog;
 
 namespace Kijk.Application;
