+++ apps/api/src/Application/Resources/Update/Request.cs
@@ -1,0 +1,3 @@
+namespace Kijk.Application.Resources.Update;
+
+public record UpdateResourceRequest(string? Name, string? Color, string? Unit);
