+++ apps/api/src/Application/Resources/Update/UpdateResource.cs
@@ -1,16 +1,14 @@
-﻿using Kijk.Application.Resources.Shared;
-using Kijk.Infrastructure.Persistence;
+﻿using Kijk.Application.Abstractions.Persistence;
+using Kijk.Application.Resources.Shared;
 using Kijk.Shared;
 using Microsoft.Extensions.Logging;
 
-namespace Kijk.Application.Resources;
-
-public record UpdateResourceRequest(string? Name, string? Color, string? Unit);
+namespace Kijk.Application.Resources.Update;
 
 /// <summary>
 /// Handler for updating a resource.
 /// </summary>
-public class UpdateResourceHandler(AppDbContext dbContext, CurrentUser currentUser, ILogger<UpdateResourceHandler> logger) : IHandler
+public class UpdateResourceHandler(IAppDbContext dbContext, CurrentUser currentUser, ILogger<UpdateResourceHandler> logger) : IHandler
 {
     public async Task<Result<ResourceResponse>> UpdateAsync(Guid id, UpdateResourceRequest request, CancellationToken cancellationToken)
     {
@@ -25,7 +23,7 @@ public class UpdateResourceHandler(AppDbContext dbContext, CurrentUser currentUs
             return Error.NotFound("User was not found");
         }
 
-        var resource = await dbContext.Resources.FindAsync([id], cancellationToken);
+        var resource = user.Resources.FirstOrDefault(x => x.Id == id);
         if (resource is null)
         {
             logger.LogWarning("Resource with id '{Id}' could not be found", id);
