+++ apps/api/src/Application/Users/Update/Request.cs
@@ -1,0 +1,3 @@
+namespace Kijk.Application.Users.Update;
+
+public record UpdateUserRequest(string? UserName, bool? UseDefaultResources);
