+++ apps/api/src/Application/Users/Welcome/Request.cs
@@ -1,0 +1,3 @@
+namespace Kijk.Application.Users.Welcome;
+
+public record WelcomeUserRequest(string? UserName, bool? UseDefaultResources);
