+++ apps/api/src/Infrastructure/Infrastructure.csproj
@@ -35,6 +35,7 @@
         <!-- Telemetry -->
     </ItemGroup>
     <ItemGroup>
+      <ProjectReference Include="..\Application\Application.csproj" />
       <ProjectReference Include="..\Domain\Domain.csproj" />
     </ItemGroup>
     <ItemGroup>
